Fix error: "range() integer end argument expected, got float." when executing tilecache_seed.py.

+++ tilecache/TileCache/Layer.py
@@ -510,8 +510,8 @@ class Layer (object):
         x0 = (minx - self.bbox[0]) / (res * self.size[0])
         y0 = (miny - self.bbox[1]) / (res * self.size[1])
         
-        x = round(x0)
-        y = round(y0)
+        x = int(round(x0))
+        y = int(round(y0))
         
         tilex = ((x * res * self.size[0]) + self.bbox[0])
         tiley = ((y * res * self.size[1]) + self.bbox[1])
